Re-add Tashk et al. 2019 (DOI: 10.1109/ICCAIRO47923.2019.00015)
When the information about the training datasets was added to
the detection and localization studies, this paper was removed
because it was not clear from the paper how they used the public
databases. After talking with the authors, the paper is added
again. They have stated that used each public dataset separately,
dividing them into train, validation and test.